MDLV 인덱스 폭이 정점 수를 따르게 한다 — u32 메시 17개가 파괴 렌더 중이었다
정본 `spec/formats/mdl-deep.json` `format.mdl.indexWidth`(확정)는 인덱스 원소 폭이 정점 수로
정해진다고 적는다: vertexCount ≤ 65535 면 u16, 넘으면 u32. 파서는 무조건 u16 이었다.

그 오독이 위험한 건 조용하기 때문이다. 정본이 증상까지 못박아 뒀다 — "u32 블롭을 u16 로 읽으면
상위 워드 0 이 섞여 maxIndex 가 정확히 0xFFFF 로 찍히고, 정점 수보다 작으므로 `maxIndex <
vertexCount` 류의 검사로는 절대 안 걸린다". `Model3D.parse` 의 그 가드가 정확히 그렇게
통과시켰고, 인덱스 개수만 2배가 되어 파스가 "성공"했다. 실패 신호가 어디에도 없었다.

도달: 코퍼스 11파일 / 17메시, camera3D 씬 7개 중 5개. 최대는 3589454154 의 陨石.mdl
(정점 3,144,456). 화면에서는 소행성·새턴 링·루시상·우주정거장·소닉 스테이지가 정점 0 을 향한
슬리버 부채꼴로 찢어진다.

저장은 u32 로 통일했다. 폭을 타입으로 들고 다니면 GPU 인덱스버퍼 바인딩까지 분기가 번지는데,
u16 메시가 941/941 중 924 로 압도적이라 그 분기의 이득이 인덱스 메모리 2배보다 작다.
GPU 드로우 7사이트도 `.uint32` 로 맞췄다(인덱스버퍼는 `Mesh.indices` 를 바이트 복사하므로
원소 폭과 드로우 폭이 어긋나면 즉시 깨진다 — 한 단위로 바꾼다).

검증:
- 신규 `Model3DIndexWidthTests` 4건 — u16 무회귀 / u32 경로 / 정본 판별식(maxIndex == vCount-1,
  0xFFFF 는 u16 오독의 지문) / 범위 밖 인덱스 거부. 수정 전에는 컴파일조차 안 된다(기대값
  69,999 가 UInt16 에 안 들어간다) — 그 자체가 결함의 증거다.
- 실물 코퍼스 전수(425 .mdl): **941/941 메시가 maxIndex == vertexCount-1**, 불일치 0.
  u32 메시 정확히 17개로 정본 실측치와 일치. (파스 nil 2건은 MDLV0013 — acceptedMagics 가
  의도적으로 제외하는 네이티브 2D 퍼펫이고 PuppetModel 담당이라 무관.)
- 전 스위트 순차 release(축소 코퍼스 38): 2,153 실행 / 실패 1 — 그 1건은 아래 사유로 무관 확인.

`PuppetModel.fromModel3D` 는 u16 유지다. 2D 퍼펫은 CPU 스키닝이고 병합 시 0xFFFF 초과 삼각형을
이미 드롭한다 — `Int(...)` 경유라 원본 폭이 넓어져도 그 방어가 그대로 성립한다.

+++ Sources/WapleRender/SceneRenderer3D.swift
@@ -1327,7 +1327,7 @@ extension SceneRenderer {
                             encoder.setFragmentTexture(mesh.texture, index: 0)
                         }
                         encoder.drawIndexedPrimitives(
-                            type: .triangle, indexCount: mesh.indexCount, indexType: .uint16,
+                            type: .triangle, indexCount: mesh.indexCount, indexType: .uint32,
                             indexBuffer: mesh.ibuf, indexBufferOffset: 0)
                     }
                 }
@@ -1615,7 +1615,7 @@ extension SceneRenderer {
                             var rp = SIMD4<Float>(mesh.refractAmount, mesh.refractRG88 ? 1 : 0, 0, 0)
                             enc.setFragmentBytes(&rp, length: MemoryLayout<SIMD4<Float>>.stride, index: 5)
                             enc.drawIndexedPrimitives(type: .triangle, indexCount: mesh.indexCount,
-                                                      indexType: .uint16, indexBuffer: mesh.ibuf, indexBufferOffset: 0)
+                                                      indexType: .uint32, indexBuffer: mesh.ibuf, indexBufferOffset: 0)
                             continue
                         }
                         // 스냅샷 확보 실패 — 재개한 enc 에서 일반 파이프라인으로 identity 폴터.
@@ -1660,7 +1660,7 @@ extension SceneRenderer {
                             var vp = viewProj
                             enc.setFragmentBytes(&vp, length: MemoryLayout<simd_float4x4>.stride, index: 6)
                             enc.drawIndexedPrimitives(type: .triangle, indexCount: mesh.indexCount,
-                                                      indexType: .uint16, indexBuffer: mesh.ibuf, indexBufferOffset: 0)
+                                                      indexType: .uint32, indexBuffer: mesh.ibuf, indexBufferOffset: 0)
                             continue
                         }
                         // 스냅샷 확보 실패 — 재개한 enc 에서 일반 파이프라인으로 identity 폴터.
@@ -1749,7 +1749,7 @@ extension SceneRenderer {
                         }
                     }
                     enc.drawIndexedPrimitives(type: .triangle, indexCount: mesh.indexCount,
-                                              indexType: .uint16, indexBuffer: mesh.ibuf, indexBufferOffset: 0)
+                                              indexType: .uint32, indexBuffer: mesh.ibuf, indexBufferOffset: 0)
                     // P⑥: 인코더 텍스처 바인딩은 다음 draw 까지 지속된다(Metal encoder 상태) — 커스텀
                     // 메시의 aux 가 slot 1(가장 흔한 textures[1])을 덮어썼다면, 같은 enc 로 뒤이어 그려질
                     // 스톡/다른 커스텀 메시(그쪽은 texture(1)을 재바인딩하지 않고 루프 진입 전 1회
